+++ src/main/java/dev/amigocode/customer/CustomerDao.java
@@ -8,10 +8,10 @@ import java.util.Optional;
 
 public interface CustomerDao {
     List<Customer> selectAllCustomers();
-    Optional<Customer> selectCustomerById(Integer id);
+    Optional<Customer> selectCustomerById(Long id);
     void insertCustomer(Customer customer);
     boolean existsPersonWithEmail(String email);
-    boolean existsPersonWithId(Integer id);
+    boolean existsPersonWithId(Long id);
     void updateCustomer(Customer update);
-    void deleteCustomerById(Integer id);
+    void deleteCustomerById(Long id);
 }
